// src/Moryx.Operators.Web/src/app/api/models/i-operator-assignable.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ICapabilities } from '../models/i-capabilities';
export interface IOperatorAssignable {
  capabilities?: ICapabilities;
  id?: number;
  name?: string | null;
  requiredSkills?: ICapabilities;
}